+++ src/main/java/frc/robot/RobotContainer.java
@@ -30,11 +30,11 @@ import frc.robot.subsystems.drive.GyroIONavX;
 import frc.robot.subsystems.drive.ModuleIO;
 import frc.robot.subsystems.drive.ModuleIOSim;
 import frc.robot.subsystems.drive.ModuleIOSpark;
-import frc.robot.subsystems.vision.Vision;
-import frc.robot.subsystems.vision.VisionConstants;
-import frc.robot.subsystems.vision.VisionIO;
-import frc.robot.subsystems.vision.VisionIOPhotonVision;
-import frc.robot.subsystems.vision.VisionIOPhotonVisionSim;
+import frc.robot.subsystems.vision.photon.Vision;
+import frc.robot.subsystems.vision.photon.VisionConstants;
+import frc.robot.subsystems.vision.photon.VisionIO;
+import frc.robot.subsystems.vision.photon.VisionIOPhotonVision;
+import frc.robot.subsystems.vision.photon.VisionIOPhotonVisionSim;
 import frc.robot.subsystems.vision.questnav.QuestNav;
 import frc.robot.subsystems.vision.questnav.QuestNavIO;
 import frc.robot.subsystems.vision.questnav.QuestNavIOReal;
